Move duplicate code into a shared function.
